Inconsistent handling of connection issues with redis
Tests are failing because GenServer processes are terminating because of
unhandeled errors. When put/get/fetch operations are executed on redis
instances that are not available, no errors are raised. We return usualy
nil values, as if the key does not exist, and log the error. Looking at
the tests, that is expected from exists? function as well, but thats not
the case. When we execute exists? function upon redis instance that cant
be reached, error is raised, and entire GenServer is terminated,
although in tests, we expect false value, as if the key does not exist.
This commit implements the logic that is expected in tests. If redis
cant be reached, error is logged, and false value is returned.